Add option to specify location of devstack config (#2784)
By default, the devstack config is now stored in a temporary folder,
which is removed when the devstack is shut down.

There are cases, however, where it is useful for the devstack config to
persist beyond the live of the devstack, allowing for inspection of the
values in the repo, or access to the databases.

By specifying `--stack-repo PATH` it is now possible for the devstack to
create a config repo in a specific location that is not removed when the
devstack closes.

+++ cmd/cli/devstack/devstack.go
@@ -11,7 +11,7 @@ import (
 	"github.com/bacalhau-project/bacalhau/cmd/util/flags/configflags"
 	"github.com/bacalhau-project/bacalhau/pkg/config"
 	"github.com/bacalhau-project/bacalhau/pkg/node"
-	"github.com/bacalhau-project/bacalhau/pkg/repo"
+	"github.com/bacalhau-project/bacalhau/pkg/setup"
 
 	"github.com/bacalhau-project/bacalhau/cmd/cli/serve"
 	"github.com/bacalhau-project/bacalhau/cmd/util"
@@ -38,6 +38,9 @@ var (
 
 		# Create a devstack cluster with a single hybrid (requester and compute) nodes
 		bacalhau devstack  --requester-nodes 0 --compute-nodes 0 --hybrid-nodes 1
+
+		# Run a devstack and create (or use) the config repo in a specific folder
+		bacalhau devstack  --stack-repo ./my-devstack-configuration
 `))
 )
 
@@ -51,6 +54,7 @@ func newDevStackOptions() *devstack.DevStackOptions {
 		CPUProfilingFile:           "",
 		MemoryProfilingFile:        "",
 		NodeInfoPublisherInterval:  node.TestNodeInfoPublishConfig,
+		ConfigurationRepo:          "",
 	}
 }
 
@@ -133,6 +137,10 @@ func NewCmd() *cobra.Command {
 		&ODs.ExecutorPlugins, "pluggable-executors", ODs.ExecutorPlugins,
 		"Will use pluggable executors when set to true",
 	)
+	devstackCmd.PersistentFlags().StringVar(
+		&ODs.ConfigurationRepo, "stack-repo", ODs.ConfigurationRepo,
+		"Folder to act as the devstack configuration repo",
+	)
 
 	return devstackCmd
 }
@@ -143,19 +151,17 @@ func runDevstack(cmd *cobra.Command, ODs *devstack.DevStackOptions, IsNoop bool)
 
 	cm := util.GetCleanupManager(ctx)
 
-	// We need to clean up the repo when the node shuts down, but we can ONLY
-	// do this because we know it is a temporary directory.
-	repoPath, _ := os.MkdirTemp("", "")
-	defer os.RemoveAll(repoPath)
+	repoPath := ODs.ConfigurationRepo
+	if repoPath == "" {
+		// We need to clean up the repo when the node shuts down, but we can ONLY
+		// do this because we know it is a temporary directory. Do not delete the
+		// configured repo if `--stack-repo` was specified
+		repoPath, _ = os.MkdirTemp("", "")
+		defer os.RemoveAll(repoPath)
+	}
 
-	fsRepo, err := repo.NewFS(repoPath)
+	fsRepo, err := setup.SetupBacalhauRepo(repoPath)
 	if err != nil {
-		return err
-	}
-	if err = fsRepo.Init(); err != nil {
-		return err
-	}
-	if err := fsRepo.Open(); err != nil {
 		return err
 	}
